Fixed edge case in Hireling set up with The Exile
This prevents The Exile from causing us to exclude too many factions, at the cost of a very rare edge case scenario which will cause the choose hireling step to randomly fail/succeed it's validation. This is due to the random order faction hirelings are chosen and is frankly not worth the massive effort and runtime processing to fix.