Added sorting to component toggles
This was required as the order you enabled expansions would change the component ordering. Left it disabled for the expansion component toggle as it's order is always determined by content.json (and we'd like to order the expansions ourselves).
Removed all "The"'s from component names so the sorting is better and we can freely preappend "the" in titles!