Added Toast for displaying errors
Changed behaviour of componentToggle so we now show an error telling you why a component is locked when you try to interact with it (this fixes the lock reason being only visible on desktop).
Added temporary images for factions and hirelings from marauders (will require updating as used images are inaccurate).
Renamed some folders in public/images so they describe physical component instead of game component.