Optimized the build, reducing chunks size & number
This involved moving the service registration and i18next initialization code into the app chunk. This fixes us needing an entire chunk just for the couple of lines of initialization code, plus a chunk just for i18next. Now, there is just 1 chunk for application code and another for dependency code.
Removed all instances of format strings in the production build. While they are a nice feature, webpack compiles them in a stupid space-inefficient way.