Added scroll shadow to component & faction select
This necessitated giving solid-colour backgrounds to the lists. The component toggle returns to it's original look when the component toggle switches to big-screen mode and enables wrapping.
Replaced beige with cornsilk as it's a more accurate darkening of ivory.
Expanded the Readme to include instructions for adding more content, and clarify more points about the translation process.
Fixed ordering error in auset-en.json