Refactored locking logic to live in redux state
This removes unnecessary duplication of lock logic between the step components and the next step thunk, and allows us to persist the user's preference for the enable/disable state of a given component between being locked and unlocked.
Slightly increased ordered list padding as the numbers were spilling over the block edge.