Initial conversion of project to Vite
This will massively improve the development experience as well as gives us flexibility that CRA didn't. Also transitioned the project to pnpm for maximum speed. Redid the PWA assets and code to be in line with the Vite standard. Rewrote code to use ES6+ features that were previously shied away from due to CRA compiling to ES5.